+++ internal/model/automigrate/base.go
@@ -14,7 +14,7 @@ import (
 )
 
 type BaseModel struct {
-	ID        int32          `gorm:"type:bigint"`
+	ID        uint64         `gorm:"type:bigint"`
 	CreatedAt time.Time      `gorm:"column:add_time"`
 	UpdatedAt time.Time      `gorm:"column:update_time"`
 	DeletedAt gorm.DeletedAt `gorm:"column:delete_time"`
